[infrt] move pd dialect position. test=develop (#40616)
