+++ paddle/infrt/dialect/pd/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(common)
+add_subdirectory(ir)
+add_subdirectory(pass)
