+++ paddle/infrt/dialect/pd/common/CMakeLists.txt
@@ -1,0 +1,4 @@
+core_gather_headers()
+
+gather_srcs(infrt_src SRCS
+    )
