+++ paddle/infrt/dialect/pd/ir/CMakeLists.txt
@@ -1,0 +1,5 @@
+core_gather_headers()
+
+gather_srcs(infrt_src SRCS
+    pd_ops.cc
+    )
