+++ paddle/infrt/dialect/pd/pass/CMakeLists.txt
@@ -1,0 +1,5 @@
+core_gather_headers()
+
+gather_srcs(infrt_src SRCS
+    pd_op_fuse_pass.cc
+    )
